Feature/delete folder with data (#52)
* fix update bug

* set up generic dialog component

* set up calls

* align end

* pop up the dialog

* lint fix

* close dialog

* dfs delete

* generic dialog test

* add test for checking modal

* lint

+++ vaulteaclient/src/app/vault/vault/vault.component.ts
@@ -1,17 +1,23 @@
+import { combineLatest, Observable, of, Subscription } from "rxjs";
+import { catchError, map, take, tap } from "rxjs/operators";
+import { EditData } from "src/app/shared/models/edit-data.interface";
+
 import { BaseComponent } from "@abstract";
-import { ChangeDetectionStrategy, Component, OnDestroy, OnInit, ViewEncapsulation } from "@angular/core";
+import {
+  ChangeDetectionStrategy, Component, OnDestroy, OnInit, ViewEncapsulation
+} from "@angular/core";
 import { MatDialogConfig } from "@angular/material/dialog";
 import { ActivatedRoute, Router } from "@angular/router";
 import { Folder, FolderFormComponent, FolderService } from "@folder";
 import { CreateItemSelectComponent, TypeEnum } from "@shared";
 import { CardData, DialogService } from "@ui-kit";
-import { combineLatest, Observable, of, Subscription } from "rxjs";
-import { catchError, map, take } from "rxjs/operators";
-import { EditData } from "src/app/shared/models/edit-data.interface";
 
 import { UserDataService } from "../../abstract/services/user-data.service";
-import { PasswordFormComponent } from "../../password/components/password-form/password-form.component";
+import {
+  PasswordFormComponent
+} from "../../password/components/password-form/password-form.component";
 import { PasswordService } from "../../password/services/password.service";
+import { GenericDialogData } from "../../ui-kit/generic-dialog/generic-dialog-data.interface";
 import { SnackBarService } from "../../ui-kit/services/snack-bar.service";
 import { VaultItem } from "../models/vault-item.interface";
 
@@ -31,6 +37,7 @@ export class VaultComponent extends BaseComponent implements OnInit, OnDestroy {
   public routeSubscription: Subscription;
 
   public editComponentMap: Map<TypeEnum, any> = new Map();
+  public vaultItems: VaultItem[] = [];
 
   public constructor(
     public userDataService: UserDataService,
@@ -77,11 +84,14 @@ export class VaultComponent extends BaseComponent implements OnInit, OnDestroy {
     );
 
     this.vaultItemsObservable = combineLatest([folderVaultObservable, passwordVaultObservable])
-      .pipe(
+    .pipe(
+        tap((result: [VaultItem[], VaultItem[]]) => {
+          this.vaultItems = [].concat(result[0], result[1]);
+        }),
         map(result => {
           return [].concat(
             result[0].filter(x => this.currentFolderId ? x.object.folderId?.toString() === this.currentFolderId : !x.object.folderId),
-            result[1].filter(x => this.currentFolderId ? x.object.folderId?.toString() === this.currentFolderId : x)
+            result[1].filter(x => this.currentFolderId ? x.object.folderId?.toString() === this.currentFolderId : !x.object.folderId)
           );
         })
     );
@@ -90,13 +100,37 @@ export class VaultComponent extends BaseComponent implements OnInit, OnDestroy {
   public ngOnInit(): void {
     this.route.params.subscribe(params => {
       this.currentFolder = this.userDataService.getFolders().find(x => x.id.toString() === params.id);
-    })
+    });
+  }
+
+  private getDeleteModalData(cardData: CardData): GenericDialogData {
+    return {
+      headerText: "Delete Folder and Contents",
+      text: `${cardData.object.name} contains other items.
+      Are you sure you want to delete ${cardData.object.name} and all of its contents?
+      This data cannot be recovered.`,
+      primaryButton: this.BUTTONS_CONSTANT.DELETE_BUTTON_DANGER,
+      secondaryButton: this.BUTTONS_CONSTANT.CANCEL_BUTTON
+    };
   }
 
   public handleDelete(cardData: CardData): void {
     switch (cardData.type) {
       case TypeEnum.FOLDER:
-        this.deleteFolder(cardData.object.id);
+        if (this.vaultItems.some(i => i.object.folderId === cardData.object.id)) {
+          this.dialogService.openWarning(this.getDeleteModalData(cardData))
+            .afterClosed()
+            .pipe(
+              take(1)
+            )
+            .subscribe(primaryButtonClicked => {
+            if (primaryButtonClicked) {
+              this.deleteFolder(cardData.object.id);
+            }
+          });
+        } else {
+          this.deleteFolder(cardData.object.id);
+        }
         break;
       case TypeEnum.PASSWORD:
         this.deletePassword(cardData.object.id);
